chore: fix tests, add feature flag for shape uploads

+++ lib/arrow/shuttle.ex
@@ -51,40 +51,70 @@ defmodule Arrow.Shuttle do
 
   """
   def create_shape(attrs \\ %{}) do
-    case upload_shape(attrs["filename"]) do
-      {:ok, new_attrs} ->
-        %Shape{}
-        |> Shape.changeset(Enum.into(new_attrs, attrs))
-        |> Repo.insert()
+    if Map.has_key?(attrs, "filename") do
+      case upload_shape(attrs["filename"]) do
+        {:ok, new_attrs} ->
+          do_create_shape(Enum.into(new_attrs, attrs))
 
-      {:error, e} ->
-        {:error, e}
+        {:error, e} ->
+          {:error, e}
+      end
+    else
+      do_create_shape(attrs)
     end
   end
 
-  defp upload_shape(%Plug.Upload{} = upload) do
-    %{bucket: bucket, prefix: prefix} = Application.get_env(:arrow, :shape_storage)
-    filename = upload.filename
-    path = "#{prefix}#{filename}"
+  defp do_create_shape(attrs) do
+    %Shape{}
+    |> Shape.changeset(attrs)
+    |> Repo.insert()
+  end
 
+  defp upload_shape(%Plug.Upload{} = upload) do
+    %{bucket: bucket, prefix: prefix, enabled?: enabled?, prefix_env: prefix_env} =
+      Application.get_env(:arrow, :shape_storage)
+
+    if enabled? do
+      prefix_env_value = System.get_env(prefix_env)
+      filename = upload.filename
+
+      path =
+        if prefix_env_value,
+          do: "#{prefix}#{prefix_env_value}/#{filename}",
+          else: "#{prefix}#{filename}"
+
+      case do_upload_shape(upload.path, bucket, path) do
+        error = {:error, _} -> error
+        {:ok, _} -> {:ok, %{"bucket" => bucket, "prefix" => prefix, "path" => path}}
+      end
+    else
+      {:ok, %{"bucket" => "disabled", "prefix" => "disabled", "path" => "disabled"}}
+    end
+  end
+
+  defp do_upload_shape(local_path, bucket, remote_path) do
     # Check if file exists already:
-    {:ok, check} = ExAws.S3.list_objects_v2(bucket, prefix: path) |> ExAws.request()
+    {:ok, check} = ExAws.S3.list_objects_v2(bucket, prefix: remote_path) |> ExAws.request()
 
     if length(check.body.contents) > 0 do
       {:error, :already_exists}
     else
       {:ok, _} =
-        upload.path
+        local_path
         |> ExAws.S3.Upload.stream_file()
-        |> ExAws.S3.upload(bucket, path)
+        |> ExAws.S3.upload(bucket, remote_path)
         |> ExAws.request()
-
-      {:ok, %{"bucket" => bucket, "prefix" => prefix, "path" => path}}
     end
   end
 
   defp delete_shape_file(shape) do
-    ExAws.S3.delete_object(shape.bucket, shape.path) |> ExAws.request()
+    %{enabled?: enabled?} = Application.get_env(:arrow, :shape_storage)
+
+    if enabled? do
+      ExAws.S3.delete_object(shape.bucket, shape.path) |> ExAws.request()
+    else
+      {:ok, :disabled}
+    end
   end
 
   @doc """
@@ -100,18 +130,25 @@ defmodule Arrow.Shuttle do
 
   """
   def update_shape(%Shape{} = shape, attrs) do
-    case upload_shape(attrs["filename"]) do
-      {:ok, new_attrs} ->
-        # Delete old file:
-        {:ok, _} = delete_shape_file(shape)
+    if Map.has_key?(attrs, "filename") do
+      case upload_shape(attrs["filename"]) do
+        {:ok, new_attrs} ->
+          # Delete old file:
+          {:ok, _} = delete_shape_file(shape)
+          do_shape_update(shape, Enum.into(new_attrs, attrs))
 
-        shape
-        |> Shape.changeset(Enum.into(new_attrs, attrs))
-        |> Repo.update()
-
-      {:error, e} ->
-        {:error, e}
+        {:error, e} ->
+          {:error, e}
+      end
+    else
+      do_shape_update(shape, attrs)
     end
+  end
+
+  defp do_shape_update(shape, attrs) do
+    shape
+    |> Shape.changeset(attrs)
+    |> Repo.update()
   end
 
   @doc """
@@ -127,8 +164,8 @@ defmodule Arrow.Shuttle do
 
   """
   def delete_shape(%Shape{} = shape) do
+    {:ok, _} = delete_shape_file(shape)
     Repo.delete(shape)
-    delete_shape_file(shape)
   end
 
   @doc """
